pyproject.toml
[project]
name = "v4"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "black>=26.1.0",
    "einops>=0.8.2",
    "gradio>=6.5.1",
    "isort>=7.0.0",
    "langchain>=1.2.8",
    "langchain-anthropic>=1.3.3",
    "langchain-chroma>=1.1.0",
    "langchain-huggingface>=1.2.0",
    "langchain-openai>=1.1.7",
    "langchain-pinecone>=0.2.13",
    "langchain-tavily>=0.2.17",
    "langchain-text-splitters>=1.1.0",
    "langgraph>=1.0.7",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.2",
    "torch>=2.10.0",
]
